requirements.txt
langchain-core>=0.3.0,<0.4.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.0,<0.4.0
langgraph>=0.4.5,<0.5.0
# faiss-cpu==1.7.4  # 제거됨 (Chroma로 대체)
chromadb>=0.4.18,<0.5.0  # 안정적인 버전 선택
langchain-chroma>=0.2.3,<0.3.0  # langchain과 chromadb를 연결하는 패키지
python-dotenv==1.0.0
typer==0.9.0
rich==13.6.0
reportlab==4.0.7
python-docx==1.0.1
markdown==3.5
bs4==0.0.1
brave-search==0.1.8
brave-search-python-client>=0.2.0
pypdf==3.17.1
docx2txt==0.8